Add router-api instructions
This commit updates the payload to the publishing-api with instructions
for the router-api.

This has become possible through the update from PR [1]. This PR make it
possible for these fields to be included and updates on the router-api.

redirect_type option has not been added because the default value
is already set for us as permanent. See [2] for more on this.

[1] https://github.com/alphagov/publishing-api/pull/874
[2] https://github.com/alphagov/gds-api-adapters/blob/master/lib/gds_api/router.rb#L33-L44